Port tests to the TCK
- Remove ported tests
- Move test class to acceptance module
- Move tests around integer overflow to bottom of class
